Added decimal support. Fixed test cases when run on Sunday.
This does intentionally change behavior where 15:20.50 is no longer 50 seconds, but instead 1/2 of a minute.  Old behavior can be enabled with: Chronic.Tags.Repeaters.RepeaterTime.OLD_TIME_DECIMAL_BEHAVIOR = true;

+++ src/Chronic/Handlers/OdRmnSyHandler.cs
@@ -13,14 +13,14 @@ namespace Chronic.Handlers
             var month = (int) tokens[1].GetTag<RepeaterMonthName>().Value;
             var year = tokens[2].GetTag<ScalarYear>().Value;
             var timeTokens = tokens.Skip(3).ToList();
-            if (Time.IsMonthOverflow(year, month, day))
+            if (Time.IsMonthOverflow((int)year, month, day))
             {
                 return null;
             }
 
             try
             {
-                var dayStart = Time.New(year, month, day);
+                var dayStart = Time.New((int)year, month, day);
                 return Utils.DayOrTime(dayStart, timeTokens, options);
             }
             catch (ArgumentException)
