+++ src/Chronic/Tags/Repeaters/RepeaterDayName.cs
@@ -66,7 +66,7 @@ namespace Chronic.Tags.Repeaters
             return base.GetNextSpan(pointer);
         }
 
-        public override Span GetOffset(Span span, int amount, Pointer.Type pointer)
+        public override Span GetOffset(Span span, decimal amount, Pointer.Type pointer)
         {
             throw new System.NotImplementedException();
         }
