+++ src/Chronic/Tags/Repeaters/RepeaterSeason.cs
@@ -23,7 +23,7 @@ namespace Chronic.Tags.Repeaters
             throw new NotImplementedException();
         }
 
-        public override Span GetOffset(Span span, int amount, Pointer.Type pointer)
+        public override Span GetOffset(Span span, decimal amount, Pointer.Type pointer)
         {
             throw new NotImplementedException();
         }
